SmartHire/requirements.txt
# Django Core
Django>=4.2,<5.0
django-crispy-forms>=2.1
crispy-bootstrap5>=2023.10
python-decouple>=3.8

# Database
# mysqlclient>=2.2.0  # Uncomment for MySQL support

# Celery for async tasks
celery>=5.3.4
django-celery-results>=2.5.1
redis>=5.0.1

# Data Science & ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Computer Vision & Video
opencv-python>=4.8.0
fer>=22.5.0

# NLP & Resume Parsing
nltk>=3.8.1
spacy>=3.7.0
pyresparser>=1.0.6

# AWS Integration
boto3>=1.28.0

# IBM Watson
ibm-watson>=7.0.0

# Others
Pillow>=10.0.0


